feat(android): ekran korpe i narucivanje (POST /api/porudzbine)
CartScreen: lista stavki (+/-/ukloni), izbor tipa porudzbine, napomena,
ukupno i dugme Naruci. CartViewModel.naruci salje PorudzbinaRequest sa
Bearer tokenom, prazni korpu i vraca na Meni.
Meni zaglavlje ima "Korpa (N)"
dugme koje vodi na korpu (nova ruta u NavHost-u)

+++ glumacplusandroid/app/src/main/java/com/example/glumac_plus_android/data/model/Porudzbina.kt
@@ -19,7 +19,7 @@ enum class StatusPorudzbine {
 @Serializable
 data class StavkaPorudzbine(
     val nazivProizvoda: String,
-    val kolicina: Int,
+    val kolicina: Double,
     val cena: Double,
     val iznosStavke: Double
 )
